+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+unreleased
+----------
+
+Added
+^^^^^
+- The methods `find_nearest_frequency()` and `find_nearest_time()` were extended with a new parameter `method`. (PR #947)
 
 0.8.0 (2026-03-16)
 ------------------